% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone-lto.R
\name{clone-lto}
\alias{clone-lto}
\alias{clone_active_themes_tbl}
\alias{clone_active_stories_tbl}
\alias{clone_active_collections_tbl}
\alias{clone_active_metadata_tbl}
\title{Clone LTO data}
\usage{
clone_active_themes_tbl()

clone_active_stories_tbl()

clone_active_collections_tbl()

clone_active_metadata_tbl()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Clone internally stored active \acronym{LTO} version data.

\code{clone_active_themes_tbl()} returns a tibble of active \acronym{LTO}
version themes.

\code{clone_active_stories_tbl()} returns a tibble of active \acronym{LTO}
version stories.

\code{clone_active_collections_tbl()} returns a tibble of active \acronym{LTO}
version collections.

\code{clone_active_metadata_tbl()} returns a tibble of active \acronym{LTO}
version metadata.
}
\examples{
\dontrun{
# Make copies of the LTO demo data:
set_lto("demo")
themes_tbl <- clone_active_themes_tbl()
stories_tbl <- clone_active_stories_tbl()
collections_tbl <- clone_active_collections_tbl()
metadata_tbl <- clone_metadatae_stories_tbl()
}
}
\seealso{
Run \link{lto-demo} to view the \acronym{LTO} demo data help page.

Run \link{lto} to find out how to load \acronym{LTO} versions.
}
