% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lto.R
\name{generate_metadata_tbl}
\alias{generate_metadata_tbl}
\title{Generate an LTO metadata tibble}
\usage{
generate_metadata_tbl(version, overwrite_rds = FALSE, verbose = TRUE)
}
\arguments{
\item{version}{A length-one character vector specifying an LTO version tag.
Set to "latest" to configure the latest numbered version, and "dev" to
configure the LTO developmental version.}

\item{overwrite_rds}{A logical value indicating whether previously
generated Rds files (if any) should be regenerated and overwritten.}

\item{verbose}{A logical value indicating whether status messages should be
output to console.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{generate_metadata_tbl()} generates (and caches) a tibble of \acronym{LTO}
metadata.
}
\details{
This function is called internally by the \code{configure_lto()}
package function. Use it directly at your own risk.
}
\keyword{internal}
