% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoRy.R
\docType{package}
\name{stoRy-package}
\alias{stoRy}
\alias{stoRy-package}
\title{stoRy: Download, Explore, and Analyze Literary Theme Ontology Data}
\description{
Download, explore, and analyze Literary Theme Ontology themes and thematically annotated story data. To learn more about the project visit \url{https://github.com/theme-ontology/theming} and \url{https://www.themeontology.org}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \pkg{stoRy} package provides utilities for working with \acronym{LTO}
data. The \acronym{LTO} is a hierarchically organized collection of
carefully defined "themes" that can be expected to arise in multiple
"stories" (i.e. works of fiction). Included in the package are functions to
download and cache \acronym{LTO} data, explore \acronym{LTO} themes and
thematically annotated stories, and analyze the thematically annotated
story data in interesting ways.

General resources:
\itemize{
\item stoRy package GitHub repository:
\url{https://github.com/theme-ontology/stoRy}
\item \acronym{LTO} project website: \url{https://www.themeontology.org}
\item \acronym{LTO} project GitHub repositories: \url{https://github.com/theme-ontology}
\item \href{https://ceur-ws.org/Vol-2518/paper-WODHSA8.pdf}{LTO conference paper}
in \emph{Proceedings of the Joint Ontology Workshops 2019 Episode V: The
Styrian Autumn of Ontology}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/theme-ontology/stoRy}
  \item \url{https://github.com/theme-ontology/theming}
  \item \url{https://www.themeontology.org/}
  \item Report bugs at \url{https://github.com/theme-ontology/stoRy/issues}
}

}
\author{
\strong{Maintainer}: Paul Sheridan \email{paul.sheridan.stats@gmail.com} (\href{https://orcid.org/0000-0002-5484-1951}{ORCID})

Authors:
\itemize{
  \item Oshan Modi
  \item Mikael Onsjö
}

}
