% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{story_ids_by_theme}
\alias{story_ids_by_theme}
\title{Story theme summary}
\format{Data frame}
\source{
Theme Ontology. See http://www.themeontology.org for more details.
}
\description{
A table in which each theme is summarised according to the stories in which it is featured.
}
