% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{story_keywords}
\alias{story_keywords}
\title{Story keywords}
\format{Data frame}
\source{
The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
}
\description{
Miscellaneous keywords for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
}
