% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{story_themes}
\alias{story_themes}
\title{Themed stories}
\format{
Data frame
}
\source{
Theme Ontology. See http://www.themeontology.org for more details.
}
\description{
Themes for the episodes of Star Trek The Original Series (tos), The Animated Series (tas), The Next Generation (tng), and Voyager (voy).
}
