% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_formatting.R
\name{text_formatting}
\alias{text_formatting}
\alias{remove_wordwrap}
\title{Parse LTO text files}
\usage{
remove_wordwrap(text)
}
\arguments{
\item{text}{A character string.}
}
\description{
Internal functions for parsing LTO raw text files.

\code{remove_wordwrap()} removes single newline characters (i.e. "\\n") from
string, but leaves double newlines (i.e. "\\n\\n").
}
\examples{
\dontrun{
# Remove word wrap from text blocks:
text <- "Themes related to people living on well beyond what
is considered to be a

normal human lifespan."
remove_wordwrap(text)
} 
}
\keyword{internal}
