% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{theme_definitions}
\alias{theme_definitions}
\title{Theme definitions}
\format{Data frame}
\source{
The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
}
\description{
A hierarchical structured vocabulary of defined themes in four domains: the human condition, society, alternate reality, and the pursuit of knowledge.
}
