% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{theme_dict}
\alias{theme_dict}
\title{Theme dictionary}
\format{Data frame}
\source{
Theme Ontology. See http://www.themeontology.org for more details.
}
\description{
A hierarchical structured vocabulary of defined themes in four domains: the human condition, society, alternate reality, and the pursuit of knowledge.
}
