% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.r
\docType{data}
\name{themed_stories}
\alias{themed_stories}
\title{Themed stories}
\format{Data frame}
\source{
The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
}
\description{
Themes for the episodes of Star Trek The Animated Series (TAS), The Original Series (TOS), and The Next Generation (TNG).
}
