% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{todata}
\alias{todata}
\title{Theme ontology}
\format{Data frame}
\source{
The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
}
\description{
A theme ontology, i.e., a hierarchical structured vocabulary of themes in four domains: the human condition, society, alternate reality, and the pursuit of knowledge.
}
