% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{tsdata}
\alias{tsdata}
\title{Story theme summary}
\format{Data frame}
\source{
The Star Trek Project. See http://http://www.thestartrekproject.net for more details.
}
\description{
A table in which each theme is summarised accoring to the stories in which it is featured.
}
